pyproject.toml
[project]
name = "zadanie-szwch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.7",
    "django-cors-headers>=4.6.0",
    "djangorestframework>=3.16.1",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.ruff.lint.isort]
force-sort-within-sections = true
split-on-trailing-comma = true

[tool.ruff.format]
quote-style = "single"
indent-style = "space"

[dependency-groups]
dev = [
    "ruff>=0.14.2",
    "requests>=2.32.0",
]

